requirements.txt
opencv-python
gradio
numpy
requests
pyjwt